+++ ios/Classes/FlutterRTCMediaStream.h
@@ -11,7 +11,10 @@
 
 -(void)getSources:(FlutterResult)result;
 
--(void)mediaStreamTrackSwitchCamera:(RTCMediaStreamTrack *)track;
+-(void)mediaStreamTrackSwitchCamera:(RTCMediaStreamTrack *)track
+                             result:(FlutterResult) result;
+
+-(void)mediaStreamTrackCaptureFrame:(RTCMediaStreamTrack *)track
+                             toPath:(NSString *) path
+                             result:(FlutterResult) result;
 @end
-
-