+++ lib/rtc_video_view.dart
@@ -10,23 +10,19 @@ enum RTCVideoViewObjectFit {
   RTCVideoViewObjectFitCover,
 }
 
-typedef void VideoRotationChangeCallback(int textureId, int rotation);
-typedef void VideoSizeChangeCallback(
-    int textureId, double width, double height);
-
 class RTCVideoRenderer {
   MethodChannel _channel = WebRTC.methodChannel();
   int _textureId;
   int _rotation = 0;
   double _width = 0.0, _height = 0.0;
-  bool _mirror;
+  bool _mirror = false;
+  double _aspectRatio = 1.0;
   MediaStream _srcObject;
   RTCVideoViewObjectFit _objectFit =
       RTCVideoViewObjectFit.RTCVideoViewObjectFitContain;
   StreamSubscription<dynamic> _eventSubscription;
-  VideoSizeChangeCallback onVideoSizeChanged;
-  VideoRotationChangeCallback onVideoRotationChanged;
-  dynamic onFirstFrameRendered;
+
+  dynamic onStateChanged;
 
   initialize() async {
     final Map<dynamic, dynamic> response =
@@ -42,15 +38,31 @@ class RTCVideoRenderer {
   double get width => _width;
 
   double get height => _height;
-  
+
   int get textureId => _textureId;
+
+  double get aspectRatio => (_width == 0 || _height == 0)
+      ? 1.0
+      : (_rotation == 90 || _rotation == 270)
+          ? _height / _width
+          : _width / _height;
+
+  bool get mirror => _mirror;
 
   set mirror(bool mirror) {
     _mirror = mirror;
+    if (this.onStateChanged != null) {
+      this.onStateChanged();
+    }
   }
+
+  RTCVideoViewObjectFit get objectFit => _objectFit;
 
   set objectFit(RTCVideoViewObjectFit objectFit) {
     _objectFit = objectFit;
+    if (this.onStateChanged != null) {
+      this.onStateChanged();
+    }
   }
 
   set srcObject(MediaStream stream) {
@@ -62,6 +74,7 @@ class RTCVideoRenderer {
   }
 
   Future<Null> dispose() async {
+    await _eventSubscription?.cancel();
     await _channel.invokeMethod(
       'videoRendererDispose',
       <String, dynamic>{'textureId': _textureId},
@@ -77,18 +90,16 @@ class RTCVideoRenderer {
     switch (map['event']) {
       case 'didTextureChangeRotation':
         _rotation = map['rotation'];
-        if (this.onVideoRotationChanged != null)
-          this.onVideoRotationChanged(_textureId, _rotation);
         break;
       case 'didTextureChangeVideoSize':
-        _width = map['width'];
-        _height = map['height'];
-        if (this.onVideoSizeChanged != null)
-          this.onVideoSizeChanged(_textureId, _width, _height);
+        _width = 0.0 + map['width'];
+        _height = 0.0 + map['height'];
         break;
       case 'didFirstFrameRendered':
-        if (this.onFirstFrameRendered != null) this.onFirstFrameRendered();
         break;
+    }
+    if (this.onStateChanged != null) {
+      this.onStateChanged();
     }
   }
 
@@ -99,70 +110,74 @@ class RTCVideoRenderer {
 }
 
 class RTCVideoView extends StatefulWidget {
-  final RTCVideoRenderer renderer;
-  RTCVideoView(this.renderer);
+  final RTCVideoRenderer _renderer;
+  RTCVideoView(this._renderer);
   @override
-  _RTCVideoViewState createState() => new _RTCVideoViewState(renderer);
+  _RTCVideoViewState createState() => new _RTCVideoViewState(_renderer);
 }
 
 class _RTCVideoViewState extends State<RTCVideoView> {
-  final RTCVideoRenderer renderer;
-  var aspectRatio = 1.0;
-
-  _RTCVideoViewState(this.renderer);
+  final RTCVideoRenderer _renderer;
+  double _aspectRatio;
+  RTCVideoViewObjectFit _objectFit;
+  bool _mirror;
+  _RTCVideoViewState(this._renderer);
 
   @override
   void initState() {
     super.initState();
     _setCallbacks();
+    _aspectRatio = _renderer.aspectRatio;
+    _mirror = _renderer.mirror;
+    _objectFit = _renderer.objectFit;
   }
 
   @override
   void deactivate() {
     super.deactivate();
-    renderer.onVideoRotationChanged = null;
-    renderer.onVideoSizeChanged = null;
-    renderer.onFirstFrameRendered = null;
+    _renderer.onStateChanged = null;
   }
 
   void _setCallbacks() {
-    renderer.onVideoRotationChanged = (int textureId, int rotation) {
+    _renderer.onStateChanged = () {
       setState(() {
-        _updateContainerSize();
-      });
-    };
-    renderer.onVideoSizeChanged = (int textureId, double width, double height) {
-      setState(() {
-        _updateContainerSize();
-      });
-    };
-    renderer.onFirstFrameRendered = () {
-      setState(() {
-        _updateContainerSize();
+        _aspectRatio = _renderer.aspectRatio;
+        _mirror = _renderer.mirror;
+        _objectFit = _renderer.objectFit;
       });
     };
   }
 
-  void _updateContainerSize() {
-    double textureWidth = 0.0, textureHeight = 0.0;
-    if (renderer.rotation == 90 || renderer.rotation == 270) {
-      textureWidth = min(renderer.width, renderer.height);
-      textureHeight = max(renderer.width, renderer.height);
-      aspectRatio = textureWidth / textureHeight;
-    } else {
-      textureWidth = max(renderer.width, renderer.height);
-      textureHeight = min(renderer.width, renderer.height);
-      aspectRatio = textureWidth / textureHeight;
-    }
+  Widget _buildVideoView(BoxConstraints constraints) {
+    return Container(
+        width: constraints.maxWidth,
+        height: constraints.maxHeight,
+        child: FittedBox(
+            fit:
+                _objectFit == RTCVideoViewObjectFit.RTCVideoViewObjectFitContain
+                    ? BoxFit.contain
+                    : BoxFit.cover,
+            child: new Center(
+                child: new SizedBox(
+                    width: constraints.maxHeight * _aspectRatio,
+                    height: constraints.maxHeight,
+                    child: new Transform(
+                        transform: Matrix4.identity()
+                          ..rotateY(_mirror ? -pi : 0.0),
+                        alignment: FractionalOffset.center,
+                        child:
+                            new Texture(textureId: _renderer._textureId))))));
   }
 
   @override
   Widget build(BuildContext context) {
-    return new Center(
-        child: (this.renderer._textureId == null || this.renderer._srcObject == null)
-            ? new Container()
-            : new AspectRatio(
-                aspectRatio: aspectRatio,
-                child: new Texture(textureId: this.renderer._textureId)));
+    bool renderVideo =
+        (_renderer._textureId != null && _renderer._srcObject != null);
+
+    return new LayoutBuilder(
+        builder: (BuildContext context, BoxConstraints constraints) {
+      return new Center(
+          child: renderVideo ? _buildVideoView(constraints) : new Container());
+    });
   }
 }