+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 ## Changelog
 --------------------------------------------
+[0.1.6] - 2019.01.31
+
+* Add getConfiguration/setConfiguration methods for Peerconnection.
+* Add object fit for RTCVideoView.
+
+[0.1.5] - 2019.03.27
+
+* Remove unnecessary parameter for getLocalDescription method.
+
+[0.1.4] - 2019.03.26
+
+* switchCamera method now returns future with isFrontCamera as result
+* Fix camera stuck in rare cases
+* Add getLocalDescription/getRemoteDescription methods
+
+[0.1.3] - 2019.03.25
+
+* Add horizontal flip (mirror) function for RTCVideoView.
+* Fixed ScreenCapture preview aspect ratio for Android.
+
+[0.1.2] - 2019.03.24
+
+* Fix compilation failure caused by invalid code.
+
+[0.1.1] - 2019.03.24
+
+* Migrated to AndroidX using Refactoring from Andoid Studio
+* Fix mediaStreamTrackSetEnable not working.
+* Fix iOS can't render video when resolution changes.
+* Some code style changes.
+
 [0.1.0] - 2019.01.21
 
 * Fix camera switch issues.